Allow options updates after stream initializes
