+++ pkg/server/metrics.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
